Load function in utils.
