+++ package.json
@@ -21,6 +21,7 @@
     "@types/react-dom": "18.0.11",
     "@unfolded/map-sdk": "^1.6.0",
     "allotment": "^1.18.1",
+    "apache-arrow": "^11.0.0",
     "autoprefixer": "10.4.14",
     "eslint": "8.38.0",
     "eslint-config-next": "13.3.0",
